Funzione "generateUpsert" spostata nel file "guiEmbedder"
