+++ ChatSQL/code/guiEmbedder.py
@@ -1,5 +1,21 @@
+import os
 import json
-#!!!!!!!!!! DA AGGIUSTARE FILE SCHEMA.JSON PER CAMPO REFERENCES
+import streamlit as st
+
+from guiUtils import getDictionaryPath, generateUpsertCommands, upsert
+
+#genera gli upsert solo quando il dizionario dati viene cambiato
+def generateUpsert(): #!!!!! DA REALIZZARE SALVATAGGIO INDEXING
+    if st.session_state.option != None: #?????? MANCA CASO ELSE
+        #effettua gli upsert relativi al dizionario appena selezionato
+        dictionary_path = getDictionaryPath(st.session_state.option)
+
+        if dictionary_path == "Error":
+            st.session_state.chat.append({"role": "assistant", "content": "Error: file path not valid"})
+        else:
+            st.session_state.upsert_commands = generateUpsertCommands(dictionary_path) #????? LA VARIABILE DEVE ESSERE GLOBALE?
+            st.session_state.emb = upsert(st.session_state.upsert_commands) #?? SERVE VAR GLOBALE!!!!!!!!!!!!!!!!!!!!!
+            st.session_state.emb.save(f"indexes/{os.path.splitext(st.session_state.option)[0]}")
 
 
 def generatePromptAdmin(emb, user_query):  # !!!!! DA REFACTORING + FARE VERSIONE USER ED ADMIN
